#14 move components to Tasks page family
